layer: Move Pipeline state methods to their own file
Somehow these non-inline method definitions escaped the first pass.
