+++ build-android/cmake/layerlib/CMakeLists.txt
@@ -66,6 +66,7 @@ add_library(VkLayer_khronos_validation SHARED
         ${SRC_DIR}/layers/cmd_buffer_state.cpp
         ${SRC_DIR}/layers/device_memory_state.cpp
         ${SRC_DIR}/layers/image_state.cpp
+        ${SRC_DIR}/layers/pipeline_state.cpp
         ${SRC_DIR}/layers/render_pass_state.cpp
         ${SRC_DIR}/layers/core_validation.cpp
         ${SRC_DIR}/layers/drawdispatch.cpp
